+++ crates/uselesskey-cli/src/srp/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod artifact_io;
+pub(crate) mod profiles;
